+++ graphql/schema/device.js
@@ -7,13 +7,20 @@ const { loadDeviceById } = require('../loaders')
 const deviceTypeDefs = gql`
     type Device {
         _id: ID!
+        name: String
         owner: User
         managers: [User]!
+    }
+
+    input DeviceInput {
+        ownerEmail: String
+        name: String
     }
 `
 
 const deviceResolvers = {
-    createDevice: async (obj, { ownerEmail }, context, info) => {
+    createDevice: async (obj, { deviceInput }, context, info) => {
+        // Permittable by admins
         if(!context.user) {
             throw new Error('User not logged in!')
         }
@@ -22,14 +29,25 @@ const deviceResolvers = {
             throw new Error('Requires admin privileges!')
         }
 
-        if(!ownerEmail) {
-            const device = new Device()
+        const device = new Device()
+
+        if(!deviceInput) {
             await device.save()
             // Admin context, so allow infinite nesting:
             return loadDeviceById(device.id)
         }
 
-        const owner = await User.findOne({email: ownerEmail})
+        if(deviceInput.name) {
+            device.name = deviceInput.name
+        }
+
+        if(!deviceInput.ownerEmail) {
+            await device.save()
+            // Admin context, so allow infinite nesting:
+            return loadDeviceById(device.id)
+        }
+
+        const owner = await User.findOne({email: deviceInput.ownerEmail})
         if(!owner) {
             throw new Error('Owner does not exist!')
         }
@@ -38,7 +56,7 @@ const deviceResolvers = {
         const session = await Device.startSession()
         session.startTransaction()
         try {
-            const device = new Device({ owner: owner._id })
+            device.owner = owner._id
             owner.devicesOwning.push(device._id)
             await owner.save()
             await device.save()
@@ -53,6 +71,7 @@ const deviceResolvers = {
 
     },
     claimDevice: async (obj, { deviceId }, context, info) => {
+        // Permittable by users
         if(!context.user) {
             throw new Error('User not logged in!')
         }
@@ -93,6 +112,30 @@ const deviceResolvers = {
             session.endSession()
             throw err
         }
+    },
+    setDeviceName: async (obj, { deviceId, name }, context, info) => {
+        // Permittable by deviceOwners and admins
+        if(!context.user) {
+            throw new Error('User not logged in!')
+        }
+        
+        const device = await Device
+            .findOne({_id: deviceId})
+            .populate('owner', '_id')
+        if(!device) {
+            throw new Error('Device does not exist!')
+        }
+
+        if(!(context.isAdmin || context.user._id === device.owner.id)) {
+            throw new Error('User not authorized!')
+        }
+
+        device.name = name
+        await device.save()
+
+        return context.isAdmin ?
+            loadDeviceById(device.id) :
+            loadDeviceById(device.id, 2)
     }
 }