+++ models/device.js
@@ -3,6 +3,7 @@ const mongoose = require('mongoose')
 const Schema = mongoose.Schema
 
 const deviceSchema = new Schema({
+  name: String,
   owner: {
     type: Schema.Types.ObjectId,
     ref: 'User'